添加新特性界面
图片盗用新浪微博的新特性图片，去除获取天气时表格刷新的动画效果

+++ Footprint/Classes/Compose/Controller/XYComposeController.m
@@ -285,7 +285,7 @@
     if (_event.text != nil && ![_event.text isEqualToString:@"(null)"]) {
         localNoti.alertBody = self.event.text;
     }else if (_event.images.count > 0){
-        localNoti.alertBody = [NSString stringWithFormat:@"有%ld张图片", _event.images.count];
+        localNoti.alertBody = [NSString stringWithFormat:@"有%ld张图片", (unsigned long)_event.images.count];
     }else if (_event.audioDuration > 0.0){
         localNoti.alertBody = @"有录音";
     }
